Fixed zero entries in currencies table bug, added validation for bigger numbers (limit to 1e20)

+++ database/migrations/2021_04_28_124057_create_conversions_table.php
@@ -18,7 +18,7 @@ class CreateConversionsTable extends Migration
             $table->timestamps();
             $table->string('original_currency');
             $table->string('destination_currency');
-            $table->float('amount');
+            $table->double('amount', 40, 2);
         });
     }
